Remove RopeUtil

+++ rope.go
@@ -4,5 +4,5 @@ type Rope interface {
 	String() string
 	Len() int
 	Index(idx int) byte
-	Depth() byte
+	Depth() int
 }
